Update compatibility field to use FLOAT data type in Taste model and related files

+++ models/taste.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
   Taste.init(
     {
       reviews: DataTypes.INTEGER,
-      compatibility: DataTypes.INTEGER,
+      compatibility: DataTypes.FLOAT,
       userId: DataTypes.INTEGER,
     },
     {
